+++ API/Entity/AppUserRole.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace API.Entity
 {
-    public class AppUserRole : IdentityUserRole<int> // it class belongs to AspNetCore
+    public class AppUserRole : IdentityUserRole<long> // it class belongs to AspNetCore
     {
         public AppUser User { get; set; }
         public AppRole Role { get; set; }
